requirements.txt
mysql-connector-python==8.3.0
numpy==1.24.3
pandas==2.2.0
setuptools==69.0.3
streamlit==1.31.0
python-dotenv==1.0.1
streamlit_extras==0.4.0
streamlit-timeline==0.0.2
streamlit-on-Hover-tabs==1.0.1
plotly==5.19.0
pytz==2023.4
extra-streamlit-components==0.1.60
Markdown==3.5.2
markdown-it-py==3.0.0
markdownlit==0.0.7
beautifulsoup4==4.12.3
requests==2.31.0